# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Rest Truckpag",
        "version": "0.0.1"
    },
    "servers": [
        {
            "url": "http://localhost/api"
        }
    ],
    "paths": {
        "/": {
            "get": {
                "tags": [
                    "API"
                ],
                "summary": "Status da API",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/products": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Listar produtos",
                "responses": {
                    "200": {
                        "description": "Lista de produtos"
                    }
                }
            },
            "post": {
                "tags": [
                    "Products"
                ],
                "summary": "Criar um novo produto",
                "responses": {
                    "201": {
                        "description": "Produto criado com sucesso"
                    }
                }
            }
        },
        "/products/{code}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Obter detalhes de um produto",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalhes do produto"
                    }
                }
            },
            "put": {
                "tags": [
                    "Products"
                ],
                "summary": "Atualizar um produto",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Produto atualizado com sucesso"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Excluir um produto",
                "parameters": [
                    {
                        "name": "code",
                        "in": "path",
                        "description": "Código do produto",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Produto excluído com sucesso"
                    }
                }
            }
        }
    }
}